+++ ExtModels/Management/Models/SchedulerTypeExt.cs
@@ -14,5 +14,6 @@ public enum SchedulerTypeExt
     PbsPro = 2,
     Slurm = 4,
     HyperQueue = 8,
-    FirecRestSlurm = 16
+    FirecRestSlurm = 16,
+    QScheduler = 32
 }