+++ HpcConnectionFramework/SchedulerAdapters/HyperQueue/Generic/HyperQueueSchedulerAdapter.cs
@@ -380,5 +380,15 @@ internal class HyperQueueSchedulerAdapter : ISchedulerAdapter
         return allHistoricalTasks;
     }
 
+    public Task<string> GetMachineArchitectureAsync(object connectorClient, HEAppE.DomainObjects.ClusterInformation.Cluster cluster, int machineId)
+    {
+        throw new NotSupportedException("GetMachineArchitecture is not supported by HyperQueue");
+    }
+
+    public Task<string> GetMachineCalibrationAsync(object connectorClient, HEAppE.DomainObjects.ClusterInformation.Cluster cluster, int machineId, string calibrationId, string endpoint)
+    {
+        throw new NotSupportedException("GetMachineCalibration is not supported by HyperQueue");
+    }
+
     #endregion
 }