+++ HpcConnectionFramework/SchedulerAdapters/LinuxLocal/Generic/LinuxLocalSchedulerAdapter.cs
@@ -539,5 +539,15 @@ public class LinuxLocalSchedulerAdapter : ISchedulerAdapter
         return allHistoricalTasks;
     }
 
+    public Task<string> GetMachineArchitectureAsync(object connectorClient, HEAppE.DomainObjects.ClusterInformation.Cluster cluster, int machineId)
+    {
+        throw new NotSupportedException("GetMachineArchitecture is not supported by LinuxLocal");
+    }
+
+    public Task<string> GetMachineCalibrationAsync(object connectorClient, HEAppE.DomainObjects.ClusterInformation.Cluster cluster, int machineId, string calibrationId, string endpoint)
+    {
+        throw new NotSupportedException("GetMachineCalibration is not supported by LinuxLocal");
+    }
+
     #endregion
 }