+++ HpcConnectionFramework/SchedulerAdapters/Slurm/Generic/SlurmSchedulerAdapter.cs
@@ -765,5 +765,15 @@ internal class SlurmSchedulerAdapter : ISchedulerAdapter
         return TaskState.Failed;
     }
 
+    public Task<string> GetMachineArchitectureAsync(object connectorClient, HEAppE.DomainObjects.ClusterInformation.Cluster cluster, int machineId)
+    {
+        throw new NotSupportedException("GetMachineArchitecture is not supported by Slurm");
+    }
+
+    public Task<string> GetMachineCalibrationAsync(object connectorClient, HEAppE.DomainObjects.ClusterInformation.Cluster cluster, int machineId, string calibrationId, string endpoint)
+    {
+        throw new NotSupportedException("GetMachineCalibration is not supported by Slurm");
+    }
+
     #endregion
 }